[tmp] move noether to noether2
time to refactor like everything (:
