--- noether/locale_.py
@@ -1,15 +1,0 @@
-"""
-Noether locale and settings.
-
-Currently not in use.
-
-Implements and handles:
- - number formatting
- - currency formatting
- - default units
-"""
-
-import locale
-locale.setlocale(locale.LC_ALL, '')
-
-# TODO: set 'conventional' UnitSystems based on location
