+++ noether2/addons/functions/trigonometry.py
