+++ noether2/display.py
