+++ noether2/unit/catalogue/unusual.py
