+++ noether2/unit/spectrum.py
